Add ability to draw fake/invisible tiles
Use the following tile modifiers:

Empty tile:
4 = fake floor
5 = fake wall
12 (palace) = fake floor (noblue)
50 (palace) = fake wall (noblue)
5 or 50 (dungeon) = fake wall with no connections
51 (dungeon) = fake wall connected to wall on the right
52 (dungeon) = fake wall connected to wall on the left
53 (dungeon) = fake wall connected to walls on both sides

Floor tile:
5 = fake wall (dungeon and palace options similar to above)
6 = invisible floor/empty
14 = invisible floor/empty + spot1/blue

Wall tile:
4 = invisible wall (display fake floor)
6 = invisible wall (display nothing/empty)
12 (palace) = invisible wall/fake floor (noblue)
14 (palace) = invisible wall/empty (spot1/blue)

Currently it is necessary to manually specify fake wall connections; this may still be somewhat buggy.
It may be an advantage to be able to manually leave a visual clue to a secret (through a strange wall pattern).
Or we may figure out a way to do the right thing automatically.